clean: rename to chart
